SQL goo --> jpg
Cleaned up stab1.sql, including working around a postgres bug in which the
implementation of generate_series bleeds state if two multi-digit series are
generated in the same query.

Added a ruby script to take the data from the query and reformat it into a ppm
which netpbm can turn into a jpg.