Added (& broke) one armed man making cereal (constraints) example
This is a copy of the code from my repo, with time sensitivity added (it used
to ignore the specified time for transitions; now it computes the cost of a
transition as the average of the cost of the start & end states times how long
it takes to make the transition.

In adding this, I apparently broke something, 'cause he now no longer picks up
his spoon.