Fixed scalling on sound_math
Straightened out the scalling issues in sound_math to yield cleaner
results.  Everything is based (for now) off an A-below-middle-C at
440.0Hz.

Also, moved the squaring outside the integral for amplitude_in (doh!)
and added a count of samples so far so that it's just a straight average
rather than a rolling average / guess.