Preliminary (but functional!) acoustic-numeric input
Added listen_for_char which waits for a period of time in which recognised
tones are heard, followed by a period in which none are, and them maps these
to a small subset of the ASCII character set.  Numbers are built by combining
I,II,III,IIII, and V as in roman numeral, but that IV = VI = 6 (which is why
we need IIII) because tones played on top of each other don't have a left-right
ordering.

Added a simple integer entry routine, with 0-9, return and backspace.