Added block of Jello to sqlgraphics image, plus cleanup
The block of jello is still rough, bu it's there.  Added alpha blending and
cleaned up coordinate transform some more, made rescalable, etc.

The computation of the projected normal (and thus the depth of intersection)
is still awful to non-existant.  The probem is, in veiw space it's non-liniar,
so (intercept,normal) form won't cut it.  Maybe do the depth in a different
space and transform the results?  Ugh.